Convert ServiceUploadsPage.js to use ES6 modules
