+++ ui/legacy.js
@@ -34,6 +34,3 @@ import '../app/assets/javascripts/student_profile/student_profile_page.jsx';
   // (unordered)
 import '../app/assets/javascripts/student_profile/page_container.jsx';
 
-// all the rest, unordered
-import '../app/assets/javascripts/service_uploads/service_uploads_page.jsx';
-
